+++ registry/lib/components/utils/url-params-clean/index.ts
@@ -4,7 +4,7 @@ import { isNotHtml, matchPattern } from '@/core/utils'
 import { useScopedConsole } from '@/core/utils/log'
 
 const displayName = '网址参数清理'
-const console = useScopedConsole({ name: displayName })
+const console = useScopedConsole(displayName)
 const entry = async () => {
   if (isNotHtml()) {
     return